Move backend relationship logic into dedicated controller
